requirements.txt
streamlit
docxtpl
python-docx
google-api-python-client
google-auth
google-auth-httplib2
pillow
pandas
